moves participant panel components to participants folder
